Fixes #167

+++ lib/Db/RecipeDb.php
@@ -46,6 +46,10 @@ class RecipeDb {
         $qb->delete('cookbook_keywords')
             ->where('recipe_id = :id');
         $qb->setParameter('id', $id, IQueryBuilder::PARAM_INT);
+        
+        $qb->delete('cookbook_categories')
+            ->where('recipe_id = :id');
+        $qb->setParameter('id', $id, IQueryBuilder::PARAM_INT);
 
         $qb->execute();
     }
@@ -58,25 +62,6 @@ class RecipeDb {
             ->where('user_id = :user')
             ->orderBy('r.name');
         $qb->setParameter('user', $userId, TYPE::STRING);
-
-        $cursor = $qb->execute();
-        $result = $cursor->fetchAll();
-        $cursor->closeCursor();
-
-        return $result;
-    }
-
-    public function findAllUncategorizedRecipes(string $userId) {
-        $qb = $this->db->getQueryBuilder();
-
-        $qb->select('r.*')
-            ->from('cookbook_keywords', 'k')
-            ->where('k.user_id = :user')
-            ->andWhere("k.name = ''")
-            ->orderBy('k.name');
-        $qb->setParameter('user', $userId, TYPE::STRING);
-
-        $qb->join('k', 'cookbook_names', 'r', 'k.recipe_id = r.recipe_id');
 
         $cursor = $qb->execute();
         $result = $cursor->fetchAll();
@@ -211,6 +196,12 @@ class RecipeDb {
         $qb->execute();
         
         $qb->delete('cookbook_keywords')
+            ->where('user_id = :user')
+            ->orWhere('user_id = :empty');
+        $qb->setParameter('user', $userId, TYPE::STRING);
+        $qb->setParameter('empty', 'empty', TYPE::STRING);
+        
+        $qb->delete('cookbook_categories')
             ->where('user_id = :user')
             ->orWhere('user_id = :empty');
         $qb->setParameter('user', $userId, TYPE::STRING);
